Bring back Big Poe Count option, lost in menu refactor (#6126)
