+++ soh/soh/CrashHandlerExt.cpp
@@ -1,4 +1,4 @@
-#include "CrashHandlerExp.h"
+#include "CrashHandlerExt.h"
 #include "variables.h"
 #include "z64.h"
 #include "z64actor.h"
