+++ soh/soh/CrashHandlerExt.h
